refactor(backend): isolate Skill repository and migration domains
